Feature/styling

+++ football-team/src/views/HomeView.vue
@@ -1,6 +1,6 @@
 <template>
-  <div class="p-3">
-    <h1>User list</h1>
+  <div class="md:p-3">
+    <h1 class="font-medium">User list</h1>
     <UserListComponent
       :userList="userList"
       :pageData="pageData"
@@ -14,9 +14,10 @@
 import UserListComponent from '@/components/UserListComponent.vue'
 import { useUserStore } from '@/stores/userStore'
 import { computed, onMounted } from 'vue'
+import { useToast } from 'primevue/usetoast'
 
 const userStore = useUserStore()
-
+const toast = useToast()
 const userList = computed(() => userStore.getUserList)
 const pageData = computed(() => userStore.getPageData)
 
@@ -27,8 +28,16 @@ const fetchUsers = (pageDetails: any) => {
   userStore.fetchUserList(pageDetails)
 }
 
-const deleteUser = (id: number) => {
-  userStore.deleteUser(id)
+const deleteUser = async (id: number) => {
+  const answer = window.confirm('Are you sure you want to delete this user?')
+  if (!answer) return false
+  await userStore.deleteUser(id)
+  toast.add({
+    severity: 'success',
+    summary: 'Delete user',
+    detail: 'User deleted successfully',
+    life: 3000
+  })
 }
 
 onMounted(() => {
